requirements.txt
keyboard==0.13.5
rich==13.7.0
pyairtable==3.0.1
python-dotenv==1.0.1
requests==2.31.0